+++ CHANGELOG.md
@@ -19,7 +19,8 @@ Changed
 - The `get_all`, `set_all`, and `delete_all` optional parameter options in `JsonParser` have
   been renamed to just `all`. This shouldn't have any effect since `JsonParser` is not intended
   to be used directly, and that option was not in `ConfigFile`.
-- `path`, `parser`, and `contents` on the `ConfigFile` are now all properties and can only be retrieved.
+- `path` on the `ConfigFile` is now a property that can only be retrieved. `contents` is now private
+  but you can use `.stringify()` instead to get the contents. `parser` is now also private.
 - `restore_original` in `ConfigFile` now raises a `FileNotFoundError` instead of an `OSError` if the original
   file path does not exist.